requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
websocket-client>=1.6.0
python-dateutil>=2.8.2
pydantic>=2.0.0
typing-extensions>=4.8.0
TA-Lib>=0.4.28
streamlit>=1.28.0
plotly>=5.17.0

